+++ cmd/box.go
@@ -21,10 +21,10 @@ var boxCmd = &cobra.Command{
 It consists of 4 seconds inhale and 4 seconds exhale and 4 seconds holding
 your breath in between.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		breathe.RunBreatheCycles(
+		breathe.RunBreathCycles(
 			"Box breathe cycle (4x4s)",
-			breathe.GenerateBreatheCycles(
-				breathe.BreatheCycle{
+			breathe.GenerateBreathCycles(
+				breathe.BreathCycle{
 					Inhale:     4000 * time.Millisecond,
 					InhaleHold: 4000 * time.Millisecond,
 					Exhale:     4000 * time.Millisecond,