+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Refactor `BreatheCycle` to `BreathCycle`
+
 ## 0.1.0
 
-Initial release
+- Initial release